refactor: localize command surface modules
