+++ src/commands/interaction/gesture.ts
@@ -8,8 +8,8 @@ import {
   optionalCliNumber,
   optionalNumber,
   requiredDaemonString,
-} from './common.ts';
-import type { CliReader, DaemonWriter, CommandInput } from './types.ts';
+} from '../cli-grammar/common.ts';
+import type { CliReader, DaemonWriter, CommandInput } from '../cli-grammar/types.ts';
 
 export const gestureCliReaders = {
   gesture: gestureInputFromCli,
